2018/12/25 name 변경 및 content_header를 activity_concert, ativityi_artist에서 content_concert, conetnt_artist 으로 이동
