+++ app/src/main/res/layout/activity_artist.xml
@@ -7,38 +7,20 @@
         android:layout_height="match_parent"
         tools:context=".activities.info.ArtistActivity">
 
-    <android.support.v4.widget.NestedScrollView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
     <!--<android.support.design.widget.AppBarLayout-->
-            <!--android:layout_height="wrap_content"-->
-            <!--android:layout_width="match_parent"-->
-            <!--android:theme="@style/AppTheme.AppBarOverlay">-->
+    <!--android:layout_height="wrap_content"-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:theme="@style/AppTheme.AppBarOverlay">-->
 
-        <!--<android.support.v7.widget.Toolbar-->
-                <!--android:id="@+id/toolbar"-->
-                <!--android:layout_width="match_parent"-->
-                <!--android:layout_height="?attr/actionBarSize"-->
-                <!--android:background="?attr/colorPrimary"-->
-                <!--app:popupTheme="@style/AppTheme.PopupOverlay"/>-->
+    <!--<android.support.v7.widget.Toolbar-->
+    <!--android:id="@+id/toolbar"-->
+    <!--android:layout_width="match_parent"-->
+    <!--android:layout_height="?attr/actionBarSize"-->
+    <!--android:background="?attr/colorPrimary"-->
+    <!--app:popupTheme="@style/AppTheme.PopupOverlay"/>-->
 
     <!--</android.support.design.widget.AppBarLayout>-->
 
-         <include
-                android:id="@+id/content_header"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                layout="@layout/content_header"
-              />
-
-            <include layout="@layout/content_artist"/>
-        </LinearLayout>
-
-    </android.support.v4.widget.NestedScrollView>
+    <include layout="@layout/content_artist"/>
 
 </android.support.design.widget.CoordinatorLayout>
