+++ app/src/main/res/layout/activity_concert.xml
@@ -7,25 +7,7 @@
         android:layout_height="match_parent"
         tools:context=".activities.info.ConcertActivity">
 
-    <android.support.v4.widget.NestedScrollView
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-        <LinearLayout
-                android:layout_width="match_parent"
-               android:layout_height="wrap_content"
-                android:orientation="vertical">
-
-            <include
-                    android:id="@+id/content_header"
-                    android:layout_width="match_parent"
-                   android:layout_height="wrap_content"
-                    layout="@layout/content_header"
-            />
 
             <include layout="@layout/content_concert"/>
-         </LinearLayout>
-
-    </android.support.v4.widget.NestedScrollView>
 
 </android.support.design.widget.CoordinatorLayout>
