+++ app/src/main/res/layout/content_artist.xml
@@ -1,39 +1,61 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+
+
+<android.support.v4.widget.NestedScrollView
         xmlns:android="http://schemas.android.com/apk/res/android"
         xmlns:tools="http://schemas.android.com/tools"
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:showIn="@layout/activity_artist"
-        tools:context=".activities.info.ArtistActivity"
-        android:orientation="vertical"
-        android:layout_marginHorizontal="16dp">
+        android:layout_height="match_parent">
 
-    <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:textSize="20sp"
-            android:textColor="#000000"
-            android:lineSpacingExtra="9sp"
-            android:textStyle="bold"
-            tools:text="다가오는 콘서트"
-    />
-    <View android:layout_width="match_parent"
-          android:layout_height="1dp"
-          android:layout_marginTop="0.7dp"
-          android:background="#999999"
-    />
-
-    <android.support.v7.widget.RecyclerView
-            android:id="@+id/recycler_view"
+    <LinearLayout
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:orientation="vertical"
-            app:layoutManager="android.support.v7.widget.LinearLayoutManager"
-    >
+            android:orientation="vertical">
 
-    </android.support.v7.widget.RecyclerView>
+        <include
+                android:id="@+id/content_header"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                layout="@layout/content_header"
+        />
 
-</LinearLayout>
+        <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                app:layout_behavior="@string/appbar_scrolling_view_behavior"
+                tools:showIn="@layout/activity_artist"
+                tools:context=".activities.info.ArtistActivity"
+                android:orientation="vertical"
+                android:layout_marginHorizontal="16dp">
+
+            <TextView
+                    android:layout_width="wrap_content"
+                    android:layout_height="wrap_content"
+                    android:textSize="20sp"
+                    android:textColor="#000000"
+                    android:lineSpacingExtra="9sp"
+                    android:textStyle="bold"
+                    tools:text="다가오는 콘서트"
+            />
+            <View android:layout_width="match_parent"
+                  android:layout_height="1dp"
+                  android:layout_marginTop="0.7dp"
+                  android:background="#999999"
+            />
+
+            <android.support.v7.widget.RecyclerView
+                    android:id="@+id/recycler_view"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:orientation="vertical"
+                    app:layoutManager="android.support.v7.widget.LinearLayoutManager"
+            >
+
+            </android.support.v7.widget.RecyclerView>
+
+        </LinearLayout>
+
+    </LinearLayout>
+
+</android.support.v4.widget.NestedScrollView>
